refactor: add separate rout 'boards'
